+++ src/moks/film.js
@@ -1,4 +1,6 @@
 import dayjs from "dayjs";
+import {getRandomInteger} from "../utils.js";
+import {getRandomIntegerDecimal} from "../utils.js";
 
 const TITLES = [
   `The Dance of Life`,
@@ -78,20 +80,6 @@ const POSTERS = [
   `the-great-flamarion.jpg`,
   `the-man-with-the-golden-arm.jpg`
 ];
-
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getRandomIntegerDecimal = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return ((lower + Math.random() * (upper - lower + 1)) - 1).toFixed(1);
-};
 
 const generateRandomItem = (array) => {
   const randomIndex = getRandomInteger(0, array.length - 1);
