+++ src/utils.js
@@ -1,0 +1,21 @@
+export const getRandomInteger = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+export const getRandomIntegerDecimal = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return ((lower + Math.random() * (upper - lower + 1)) - 1).toFixed(1);
+};
+
+export const limitDesc = (descriptionText, maxSymbols) => {
+  if (descriptionText.length > maxSymbols) {
+    return `${descriptionText.substring(0, maxSymbols - 1)}...`;
+  }
+
+  return descriptionText;
+};
